% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarities_calculation.R
\name{cosine_sim}
\alias{cosine_sim}
\title{Load directory where patient vcf files are in}
\usage{
cosine_sim(directory)
}
\arguments{
\item{directory}{The directory which includes patients' vcf files.}
}
\value{
30 consine similarities for each patient
}
\description{
Calculate cosine similarities with COSMIC muational signatures version 2.
}
