% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_manipulation.R
\name{feature_manipulation}
\alias{feature_manipulation}
\title{Load a data frame}
\usage{
feature_manipulation(dataframe)
}
\arguments{
\item{dataframe}{TO output dataframe to be loaded}
}
\value{
A dataframe after feature manipulation
}
\description{
feature manipulation
}
