% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prediction.R
\name{make_prediction}
\alias{make_prediction}
\title{Load a data frame}
\usage{
make_prediction(dataframe, package_path)
}
\arguments{
\item{dataframe}{TO output dataframe to be loaded}

\item{package_path}{The path of TnToClassifier package you have downloaded.}
}
\value{
A prediction dataframe.
}
\description{
Make predictions for TO_output using GBM model.
}
