% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definition.R
\name{original_data}
\alias{original_data}
\title{Load directory where TO_output and path of TN_output are in.}
\usage{
original_data(directory)
}
\arguments{
\item{directory}{The directory includes TO_output and TN_output folders. Also, the folder name of TO_output should be
"TO_output". The folder name of TN_output should be "TN_output".}
}
\value{
A dataframe including different classes(TP, FP, FN), datatype, dataset and SampleID.
}
\description{
Generate a dataframe including different classes (TP, FP, FN).
}
\examples{
original_data(sprintf("\%s/avelumab-001-personalis-TNvTO_xome", Sys.getenv("GITCLONES")))
}
