% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_nucleotide_frequency_calculation.R
\name{single_nucleotide_change_frequency}
\alias{single_nucleotide_change_frequency}
\title{Load a data frame}
\usage{
single_nucleotide_change_frequency(dataframe)
}
\arguments{
\item{dataframe}{TO output dataframe.}
}
\value{
Dataframe includes single nucleotide change frequency for each patient.
}
\description{
Calculate single nucleotide change frequence for each patient based on single nucleotide change
}
