% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definition.R
\name{TO_generate}
\alias{TO_generate}
\title{Load directory where TO_output is in. (use this function is only TO_output is available)}
\usage{
TO_generate(directory)
}
\arguments{
\item{directory}{The directory includes TO_output folder. Also, the folder name of TO_output should be
"TO_output".}
}
\value{
A dataframe including different classes, datatype, dataset and SampleID.
}
\description{
Generate a dataframe including different classes.
}
\examples{
original_data(sprintf("\%s/avelumab-001-personalis-TNvTO_xome", Sys.getenv("GITCLONES")))
}
