% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplet_and_triplet_frequency_calculation.R
\name{triplet}
\alias{triplet}
\title{Load a data frame}
\usage{
triplet(dataframe)
}
\arguments{
\item{dataframe}{TO output dataframe to be loaded}
}
\value{
A triplet change for variant.
}
\description{
Calculate triplet based on variant's chromosome sequence and position.
}
