% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplet_and_triplet_frequency_calculation.R
\name{triplet_frequency}
\alias{triplet_frequency}
\title{Load a data frame}
\usage{
triplet_frequency(dataframe)
}
\arguments{
\item{dataframe}{TO output dataframe after calculating triplet change.}
}
\value{
Triplet frequency change for each patient.
}
\description{
Calculate triplet change frequence for each patient based on triplet change
}
