% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prediction.R
\name{make_prediction}
\alias{make_prediction}
\title{Load a data frame}
\usage{
make_prediction(dataframe)
}
\arguments{
\item{dataframe}{TO output dataframe to be loaded}
}
\value{
A prediction dataframe.
}
\description{
Make predictions for TO_output using GBM model.
}
